added test for problem 73

+++ problems/Problem1.Answer/Problem1Answer.cs
@@ -2,7 +2,7 @@
 
 namespace Problem1.Answer {
   public static class Problem1Answer {
-    public static int SumMultiplesOfThreeOrFiveLessThan(int cap) =>
+    public static long SumMultiplesOfThreeOrFiveLessThan(long cap) =>
       Enumerable.Range(0, cap)
         .Where(i => i % 3 == 0 || i % 5 == 0)
         .Sum();
